apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pandas==2.2.2
polars==1.9.0
networkx==3.3
orjson==3.10.7
python-multipart==0.0.6
httpx==0.27.2
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.21.1
mypy==1.7.1
ruff==0.1.6
black==23.11.0
requests==2.31.0
